Moved search query for rental information into a rentals controller with a separate rentals route.
