+++ app.js
@@ -27,7 +27,9 @@ app.use('/customers', customers);
 
 var movies = require('./routes/movies');
 app.use('/movies', movies);
-app.use('/movie', movies);
+
+var rentals = require('./routes/rentals');
+app.use('/rentals', rentals);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
